+++ DigitalStore.Api/DigitalStore.Data/Domain/ShoppingCartItem.cs
@@ -15,6 +15,6 @@ namespace DigitalStore.Data.Domain
         public ShoppingCart ShoppingCart { get; set; }
         public int Stock { get; set; }
         public int Quantity { get; set; } // Adet
-        public decimal Price { get; set; } // Adet
+        public float Price { get; set; } // Adet
     }
 }
