+++ DigitalStore.Api/DigitalStore.Schema/OrderDetail.cs
@@ -20,8 +20,8 @@ namespace DigitalStore.Schema
         public long Id { get; set; }
 
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal TotalPrice { get; set; }
+        public float UnitPrice { get; set; }
+        public float TotalPrice { get; set; }
         public int ProductId { get; set; }
         public int OrderId { get; set; }
     }
